+++ astropy/wcs/__init__.py
@@ -1,8 +1,8 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """
 .. _wcslib: http://www.atnf.csiro.au/~mcalabre/WCS/
-.. _WCS papers: http://www.atnf.csiro.au/people/mcalabre/WCS/index.html
-.. _Paper IV: http://www.atnf.csiro.au/people/mcalabre/WCS/dcs_20040422.pdf
+.. _FITS WCS standard: http://www.atnf.csiro.au/people/mcalabre/WCS/index.html
+.. _distortion paper: http://www.atnf.csiro.au/people/mcalabre/WCS/dcs_20040422.pdf
 .. _SIP: http://irsa.ipac.caltech.edu/data/SPITZER/docs/files/spitzer/shupeADASS.pdf
 
 `astropy.wcs` contains utilities for managing World Coordinate System
@@ -12,11 +12,11 @@ position on the sky sphere.
 
 It performs three separate classes of WCS transformations:
 
-- Core WCS, as defined in the `WCS papers`_, based on Mark
+- Core WCS, as defined in the `FITS WCS standard`_, based on Mark
   Calabretta's `wcslib`_.  See `~astropy.wcs.Wcsprm`.
 - Simple Imaging Polynomial (`SIP`_) convention.  See
   `~astropy.wcs.Sip`.
-- table lookup distortions as defined in WCS `Paper IV`_.  See
+- table lookup distortions as defined in WCS `distortion paper`_.  See
   `~astropy.wcs.DistortionLookupTable`.
 
 Each of these transformations can be used independently or together in